requirements.txt
cloud-sql-python-connector[pg8000]==1.10.0
elevenlabs==1.3.1
flask-bootstrap==3.3.7.1
flask==3.0.3
google-cloud-storage==2.17.0
gunicorn==22.0.0
langchain-openai==0.1.13
langchain==0.2.6
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
SQLAlchemy==2.0.7
Werkzeug==3.0.3
